Create noofleftlower
N people are standing in line numbered 1 to N from left to right. Each person wants to know the height of the person to left of him having height less than him. If there are multiple such people he wants to know the height of the rightmost such person. If there is no such person report -1.